fix: use navigate to go to dashboard component
fix the issue of co
mpany and user data not found. the trick is to depend both on the token (i.e., token usetate variable). when we add new value to the to
ken variable, the useeffect runs i.e., the useeffect dependent on the token. so when we navigate tot the dashboard, the fetching of com
pany and user info happens in the background and then populates the component once complete