chore: add vercel.json
add rewrite route  to redirect to the the index.html whenever the pages reload. This solves the common problme of SPA, where when we hit reload the app tries to fetch a file for that path, but that's not how a SPA works. So we add a rewrite rule to fetch the index.html that the react serves